Minitalk/bonus

+++ server_bonus.c
@@ -12,21 +12,17 @@
 
 #include "minitalk.h"
 
-void	send_character(int *len, char *letter)
-{
-	if (*len == 8)
-	{
-		if (*letter == '\0')
-		{
-			write(1, "\n", 2);
-			*len = 0;
-			*letter = 0;
-		}
-		write(1, letter, 1);
-		*len = 0;
-		*letter = 0;
-	}
-}
+// void	send_character(int *len, char *letter, int *pid)
+// {
+// 	if (*len == 8)
+// 	{
+// 		if (*letter == '\0')
+// 			kill(*pid, SIGUSR2);
+// 		write(1, letter, 1);
+// 		*len = 0;
+// 		*letter = 0;
+// 	}
+// }
 
 void	handler(int sig_nbr, siginfo_t *info, void *context)
 {
@@ -34,7 +30,6 @@ void	handler(int sig_nbr, siginfo_t *info, void *context)
 	static int	bit;
 	static int	len;
 	static int	pid;
-	int			error;
 
 	(void)context;
 	if (!pid)
@@ -51,10 +46,15 @@ void	handler(int sig_nbr, siginfo_t *info, void *context)
 		bit = 1;
 	letter = (letter << 1) | (bit & 1);
 	len++;
-	send_character(&len, &letter);
-	error = kill(pid, SIGUSR1);
-	if (error == -1)
-		print_error();
+	if (len == 8)
+	{
+		if (letter == '\0')
+			kill(pid, SIGUSR1);
+		write(1, letter, 1);
+		len = 0;
+		letter = 0;
+	}
+	kill(pid, SIGUSR2);
 }
 
 int	main(void)
@@ -67,6 +67,7 @@ int	main(void)
 	write(1, "\n", 1);
 	move.sa_sigaction = handler;
 	move.sa_flags = SA_SIGINFO;
+	sigemptyset(&sa.sa_mask);
 	sigaction(SIGUSR1, &move, NULL);
 	sigaction(SIGUSR2, &move, NULL);
 	while (1)
